agents/project-updates-extractor/requirements.txt
strands-agents==1.24.0
strands-agents-tools==0.2.17
bedrock-agentcore==1.1.1
boto3>=1.42.7
requests>=2.31.0
mcp>=1.0.0
